Move fixture into dedicated location.
